% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CenteredCusumValues}
\alias{CenteredCusumValues}
\title{Centered Cusum of Values}
\usage{
CenteredCusumValues(y)
}
\arguments{
\item{y}{Vector of centered, cumulative sum of squares}
}
\value{
Vector of centered, cumulative sum of squares
}
\description{
Centered Cusum of Values
}
