% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_critical_value}
\alias{check_critical_value}
\title{Check Critical Value}
\usage{
check_critical_value(Dk)
}
\arguments{
\item{Dk}{Vector of centered, cumulative sum of squares}
}
\value{
List with position and an indicator of significance
}
\description{
Check Critical Value
}
