% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basta.R
\name{detection}
\alias{detection}
\title{BASTA Algorithm to Detect Multiple-Change-Points}
\usage{
detection(x, order = 1, factor = 8, c = 0.5, epsilon = 0, log = TRUE)
}
\arguments{
\item{x}{a numeric vector giving the time series to analyse}

\item{order}{order of the ARCH model}

\item{factor}{dampening factor \code{F} from the paper}

\item{c}{threshold constant \code{c} from the paper}

\item{epsilon}{epsilon from the paper}

\item{log}{logical, if true, then \eqn{U^{(4)}_t} is used; otherwise \eqn{U^{(3)}_t}}
}
\value{
Estimated expectation of the function \eqn{g(.)} from the paper.
}
\description{
If \code{x} is a piecewise stationary ARCH process, then \code{detection} returns the estimated
expectation of the function \eqn{g(.)} from the paper.
}
