% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICSS.R
\name{ICSS}
\alias{ICSS}
\title{Iterative Cumulative Sum of Squares (ICSS)}
\usage{
ICSS(data = data, demean = FALSE)
}
\arguments{
\item{data}{A numerical vector}

\item{demean}{An object of class "logical": If demean is TRUE, all data will get demeaned prior to the ICSS algorithm.}
}
\value{
ICSS returns a numerical vector containing the location of structural breakpoints or NA if none breakpoints are found.
}
\description{
ICSS implements the Iterative Cumulative Sum of Squares (ICSS) algorithm by Inclan/Tiao (1994).
The test detects structural breakpoints in the variance of time series data.
}
