% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_converged}
\alias{is_converged}
\title{Check Convergence}
\usage{
is_converged(oldVec, newVec)
}
\arguments{
\item{oldVec}{Integer vector of old positions.}

\item{newVec}{Integer vector of new positions.}
}
\value{
Boolean that indicates if ICSS algorithm converged.
}
\description{
Check Convergence
}
