% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basta.R
\name{xreg_basta}
\alias{xreg_basta}
\title{Create \code{xreg} For Structural Breaks Models}
\usage{
xreg_basta(y, demean = FALSE, pulse = TRUE, ...)
}
\arguments{
\item{y}{dependent variable}

\item{demean}{logical, if \code{TRUE} then \code{y} will get demeaned prior to the BASTA algorithm}

\item{pulse}{logical, if \code{TRUE} then dummy variables equal one at detected change-point. If \code{FALSE} then dummy variables equal one at and after detected breakpoint.}

\item{...}{additional arguments to pass to \code{detection}}
}
\value{
A matrix
}
\description{
Create a matrix of dummy variables for use as external regressons.
}
